# publish rework + reconnect retry buffer + inline subscriptions