--- a/AsyncNats.Tests/Messages/NatsHPub_SerializeShould .cs
+++ b/AsyncNats.Tests/Messages/NatsHPub_SerializeShould .cs
@@ -17,8 +17,8 @@
         {
             NatsMsgHeaders headers = new Dictionary<string, string>() { ["key"] = "value" };
 
-            using var rented = NatsHPub.RentedSerialize(new NatsMemoryPool(), "FOO", NatsKey.Empty, headers, Encoding.UTF8.GetBytes("Hello NATS!"));
-            var text = Encoding.UTF8.GetString(rented.Memory.Span);
+            var rented = NatsHPub.Serialize( "FOO", NatsKey.Empty, headers, Encoding.UTF8.GetBytes("Hello NATS!"));
+            var text = Encoding.UTF8.GetString(rented.Span);
 
             Assert.Equal("HPUB FOO 23 34\r\nNATS/1.0\r\nkey:value\r\n\r\nHello NATS!\r\n", text);
         }
@@ -27,8 +27,8 @@
         public void BeSameWithReplyTo()
         {
             NatsMsgHeaders headers = new Dictionary<string, string>() { ["key"] = "value" };
-            using var rented = NatsHPub.RentedSerialize(new NatsMemoryPool(), "FRONT.DOOR", "INBOX.22", headers, Encoding.UTF8.GetBytes("Knock Knock"));
-            var text = Encoding.UTF8.GetString(rented.Memory.Span);
+            var rented = NatsHPub.Serialize("FRONT.DOOR", "INBOX.22", headers, Encoding.UTF8.GetBytes("Knock Knock"));
+            var text = Encoding.UTF8.GetString(rented.Span);
 
             Assert.Equal("HPUB FRONT.DOOR INBOX.22 23 34\r\nNATS/1.0\r\nkey:value\r\n\r\nKnock Knock\r\n", text);
         }