--- a/AsyncNats.Tests/Messages/NatsMsg_ParseMessageShould.cs
+++ b/AsyncNats.Tests/Messages/NatsMsg_ParseMessageShould.cs
@@ -17,7 +17,7 @@
         {
             var reader = new SequenceReader<byte>(new ReadOnlySequence<byte>(Encoding.UTF8.GetBytes("Hello World\r\n")));
             var message = new ReadOnlySpan<byte>(Encoding.UTF8.GetBytes("MSG FOO.BAR 9 11"));
-            var msg = NatsMsg.ParseMessage(new NatsMemoryPool(), message, ref reader);
+            var msg = new NatsMessageParser().ParseMessage( message, ref reader,out _);
             Assert.IsType<NatsMsg>(msg);
             ((NatsMsg) msg).Release();
         }
@@ -27,7 +27,7 @@
         {
             var reader = new SequenceReader<byte>();
             var message = new ReadOnlySpan<byte>(Encoding.UTF8.GetBytes("MSG FOO.BAR 9 11"));
-            var msg = NatsMsg.ParseMessage(new NatsMemoryPool(), message, ref reader);
+            var msg = new NatsMessageParser().ParseMessage(message, ref reader,out _);
             Assert.Null(msg);
         }
 
@@ -36,7 +36,7 @@
         {
             var reader = new SequenceReader<byte>(new ReadOnlySequence<byte>(Encoding.UTF8.GetBytes("Hello World\r\n")));
             var message = new ReadOnlySpan<byte>(Encoding.UTF8.GetBytes("MSG FOO.BAR 9 11"));
-            var msg = (NatsMsg) NatsMsg.ParseMessage(new NatsMemoryPool(), message, ref reader);
+            var msg = (NatsMsg)new NatsMessageParser().ParseMessage(message, ref reader,out _);
             Assert.Equal("FOO.BAR", msg.Subject.AsString());
             Assert.Equal(9, msg.SubscriptionId);
             Assert.Equal(11, msg.Payload.Length);
@@ -49,7 +49,7 @@
         {
             var reader = new SequenceReader<byte>(new ReadOnlySequence<byte>(Encoding.UTF8.GetBytes("Hello World\r\n")));
             var message = new ReadOnlySpan<byte>(Encoding.UTF8.GetBytes("MSG FOO.BAR 9 INBOX.34 11"));
-            var msg = (NatsMsg) NatsMsg.ParseMessage(new NatsMemoryPool(), message, ref reader);
+            var msg = (NatsMsg)new NatsMessageParser().ParseMessage(message, ref reader,out _);
             Assert.Equal("FOO.BAR", msg.Subject.AsString());
             Assert.Equal(9, msg.SubscriptionId);
             Assert.Equal("INBOX.34", msg.ReplyTo.AsString());
@@ -63,7 +63,7 @@
         {
             var reader = new SequenceReader<byte>(new ReadOnlySequence<byte>(Encoding.UTF8.GetBytes("NATS/1.0\r\nkey:value\r\n\r\nHello World\r\n")));
             var message = new ReadOnlySpan<byte>(Encoding.UTF8.GetBytes("HMSG FOO.BAR 9 INBOX.34 23 34"));
-            var msg = (NatsMsg)NatsMsg.ParseMessageWithHeader(new NatsMemoryPool(), message, ref reader);
+            var msg = (NatsMsg)new NatsMessageParser().ParseMessageWithHeader( message, ref reader,out _);
             Assert.Equal("FOO.BAR", msg.Subject.AsString());
             Assert.Equal(9, msg.SubscriptionId);
             Assert.Equal("INBOX.34", msg.ReplyTo.AsString());
@@ -82,7 +82,7 @@
         {
             var reader = new SequenceReader<byte>(new ReadOnlySequence<byte>(Encoding.UTF8.GetBytes("NATS/1.0\r\nkey:value\r\n\r\nHello World\r\n")));
             var message = new ReadOnlySpan<byte>(Encoding.UTF8.GetBytes("HMSG FOO.BAR 9 23 34"));
-            var msg = (NatsMsg)NatsMsg.ParseMessageWithHeader(new NatsMemoryPool(), message, ref reader);
+            var msg = (NatsMsg)new NatsMessageParser().ParseMessageWithHeader(message, ref reader,out _);
             Assert.Equal("FOO.BAR", msg.Subject.AsString());
             Assert.Equal(9, msg.SubscriptionId);
             Assert.True(msg.ReplyTo.IsEmpty);