--- a/AsyncNats.Tests/Messages/NatsPong_ParseMessageShould.cs
+++ b/AsyncNats.Tests/Messages/NatsPong_ParseMessageShould.cs
@@ -14,8 +14,7 @@
         [Fact]
         public void ReturnNatsPong()
         {
-            var reader = new SequenceReader<byte>();
-            var pong = NatsPong.ParseMessage(new NatsMemoryPool(),  _message.Span, ref reader);
+            var pong = new NatsMessageParser().ParsePong();
             Assert.IsType<NatsPong>(pong);
         }
     }