--- a/AsyncNats.Tests/Messages/NatsPong_SerializeShould.cs
+++ b/AsyncNats.Tests/Messages/NatsPong_SerializeShould.cs
@@ -14,8 +14,8 @@
         [Fact]
         public void BeSame()
         {
-            using var rented = NatsPong.RentedSerialize(new NatsMemoryPool());
-            var text = Encoding.UTF8.GetString(rented.Memory.Span);
+            var rented = NatsPong.Serialize();
+            var text = Encoding.UTF8.GetString(rented.Span);
 
             Assert.Equal("PONG\r\n", text);
         }