--- a/AsyncNats.Tests/Messages/NatsSub_SerializeShould.cs
+++ b/AsyncNats.Tests/Messages/NatsSub_SerializeShould.cs
@@ -14,8 +14,8 @@
         [Fact]
         public void BeSameWithoutQueueGroup()
         {
-            using var rented = NatsSub.RentedSerialize(new NatsMemoryPool(), "FOO", NatsKey.Empty, 1);
-            var text = Encoding.UTF8.GetString(rented.Memory.Span);
+            var rented = NatsSub.Serialize("FOO", NatsKey.Empty, 1);
+            var text = Encoding.UTF8.GetString(rented.Span);
 
             Assert.Equal("SUB FOO 1\r\n", text);
         }
@@ -23,8 +23,8 @@
         [Fact]
         public void BeSameWithQueueGroup()
         {
-            using var rented = NatsSub.RentedSerialize(new NatsMemoryPool(), "BAR", "G1", 44);
-            var text = Encoding.UTF8.GetString(rented.Memory.Span);
+            var rented = NatsSub.Serialize("BAR", "G1", 44);
+            var text = Encoding.UTF8.GetString(rented.Span);
 
             Assert.Equal("SUB BAR G1 44\r\n", text);
         }