--- a/AsyncNats/AsyncNats.csproj
+++ b/AsyncNats/AsyncNats.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
-    <LangVersion>8.0</LangVersion>
+    <LangVersion>9.0</LangVersion>
     <Nullable>enable</Nullable>
     <AssemblyName>EightyDecibel.AsyncNats</AssemblyName>
     <RootNamespace>EightyDecibel.AsyncNats</RootNamespace>