--- a/AsyncNats/Messages/INatsClientMessage.cs
+++ b/AsyncNats/Messages/INatsClientMessage.cs
@@ -1,6 +1,16 @@
-﻿namespace EightyDecibel.AsyncNats.Messages
+﻿using System;
+using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
+using System.Threading.Tasks.Sources;
+
+namespace EightyDecibel.AsyncNats.Messages
 {
     public interface INatsClientMessage
     {
+        int Length { get; }
+        void Serialize(Span<byte> buffer);
+
     }
+
 }