--- a/AsyncNats/Messages/NatsError.cs
+++ b/AsyncNats/Messages/NatsError.cs
@@ -11,11 +11,6 @@
 
         public string? Error { get; set; }
 
-        public static INatsServerMessage? ParseMessage(NatsMemoryPool pool, in ReadOnlySpan<byte> line, ref SequenceReader<byte> reader)
-        {
-            if (line.Length == 6) return new NatsError();
-            var error = line.Slice(6, line.Length - 9); // Remove "-ERR ''"
-            return new NatsError {Error = Encoding.UTF8.GetString(error)};
-        }
+        
     }
 }