--- a/AsyncNats/Messages/NatsHPub.cs
+++ b/AsyncNats/Messages/NatsHPub.cs
@@ -5,22 +5,36 @@
     using System.Buffers.Text;
     using System.Collections.Generic;
     using System.Text;
+    using System.Xml.Linq;
 
 
-    public class NatsHPub
+    public readonly struct NatsHPub:INatsClientMessage
     {
         private static readonly ReadOnlyMemory<byte> _command = new ReadOnlyMemory<byte>(Encoding.UTF8.GetBytes("HPUB "));
         private static readonly ReadOnlyMemory<byte> _del = new ReadOnlyMemory<byte>(Encoding.UTF8.GetBytes(" "));
         private static readonly ReadOnlyMemory<byte> _end = new ReadOnlyMemory<byte>(Encoding.UTF8.GetBytes("\r\n"));
+                
+        public  int Length => _length;
 
-        public static IMemoryOwner<byte> RentedSerialize(NatsMemoryPool pool, in NatsKey subject, in NatsKey replyTo, in NatsMsgHeaders header, in NatsPayload payload)
+        private readonly int _length;
+        private readonly NatsKey _subject;
+        private readonly NatsKey _replyTo;
+        private readonly NatsMsgHeaders _header;
+        private readonly NatsPayload _payload;
+
+        public NatsHPub(in NatsKey subject, in NatsKey replyTo, in NatsMsgHeaders header, in NatsPayload payload)
         {
+            _subject = subject;
+            _replyTo = replyTo;
+            _header = header;
+            _payload = payload;
+
             var totalLength = payload.Memory.Length + header.SerializedLength;
-            
+
             var hint = _command.Length; // HPUB
             hint += subject.Memory.Length + 2; // Subject + spaces
             hint += replyTo.IsEmpty ? 0 : replyTo.Memory.Length + 1; // ReplyTo
-            
+
             if (header.SerializedLength < 10) hint += 1;
             else if (header.SerializedLength < 100) hint += 2;
             else if (header.SerializedLength < 1_000) hint += 3;
@@ -29,7 +43,7 @@
             else if (header.SerializedLength < 1_000_000) hint += 6;
             else if (header.SerializedLength < 10_000_000) hint += 7;
             else throw new ArgumentOutOfRangeException(nameof(header));
-            
+
             if (totalLength < 10) hint += 1;
             else if (totalLength < 100) hint += 2;
             else if (totalLength < 1_000) hint += 3;
@@ -38,52 +52,65 @@
             else if (totalLength < 1_000_000) hint += 6;
             else if (totalLength < 10_000_000) hint += 7;
             else throw new ArgumentOutOfRangeException(nameof(payload));
-            
-            
+
+
 
             hint += _end.Length; // Ending
             hint += payload.Memory.Length;
             hint += header.SerializedLength;
             hint += _end.Length; // Ending Payload
 
-            var rented = pool.Rent(hint);
-            var buffer = rented.Memory;
+            _length = hint;
+        }
 
-            _command.CopyTo(buffer);
+        public void Serialize(Span<byte> buffer)
+        {
+          
+
+            _command.Span.CopyTo(buffer);
             var consumed = _command.Length;
-            subject.Memory.Span.CopyTo(buffer.Slice(consumed).Span);
-            consumed += subject.Memory.Length;
-            _del.CopyTo(buffer.Slice(consumed));
+            _subject.Memory.Span.CopyTo(buffer.Slice(consumed));
+            consumed += _subject.Memory.Length;
+            _del.Span.CopyTo(buffer.Slice(consumed));
             consumed++;
-            if (!replyTo.IsEmpty)
+            if (!_replyTo.IsEmpty)
             {
-                replyTo.Memory.Span.CopyTo(buffer.Slice(consumed).Span);
-                consumed += replyTo.Memory.Length;
-                _del.CopyTo(buffer.Slice(consumed));
+                _replyTo.Memory.Span.CopyTo(buffer.Slice(consumed));
+                consumed += _replyTo.Memory.Length;
+                _del.Span.CopyTo(buffer.Slice(consumed));
                 consumed++;
             }
 
-            Utf8Formatter.TryFormat(header.SerializedLength, buffer.Slice(consumed).Span, out var written);
+            Utf8Formatter.TryFormat(_header.SerializedLength, buffer.Slice(consumed), out var written);
             consumed += written;
-            _del.CopyTo(buffer.Slice(consumed));
+            _del.Span.CopyTo(buffer.Slice(consumed));
             consumed++;
 
-            Utf8Formatter.TryFormat(payload.Memory.Length + header.SerializedLength, buffer.Slice(consumed).Span, out written);            
+            Utf8Formatter.TryFormat(_payload.Memory.Length + _header.SerializedLength, buffer.Slice(consumed), out written);
             consumed += written;
-            _end.CopyTo(buffer.Slice(consumed));
+            _end.Span.CopyTo(buffer.Slice(consumed));
             consumed += _end.Length;
 
-            header.SerializeTo(buffer.Slice(consumed).Span);
-            consumed += header.SerializedLength;
+            _header.SerializeTo(buffer.Slice(consumed));
+            consumed += _header.SerializedLength;
 
-            if (!payload.IsEmpty)
+            if (!_payload.IsEmpty)
             {
-                payload.Memory.CopyTo(buffer.Slice(consumed));
-                consumed += payload.Memory.Length;
+                _payload.Memory.Span.CopyTo(buffer.Slice(consumed));
+                consumed += _payload.Memory.Length;
             }
 
-            _end.CopyTo(buffer.Slice(consumed));
-            return rented;
+            _end.Span.CopyTo(buffer.Slice(consumed));
+        }
+
+        //test and debug only
+        public static ReadOnlyMemory<byte> Serialize(in NatsKey subject, in NatsKey replyTo, in NatsMsgHeaders header,in NatsPayload payload)
+        {
+            var pub = new NatsHPub(subject, replyTo, header,payload);
+            var buffer = new byte[pub.Length];
+            pub.Serialize(buffer);
+
+            return buffer;
         }
     }
 }