--- a/AsyncNats/Messages/NatsMsgHeaders.cs
+++ b/AsyncNats/Messages/NatsMsgHeaders.cs
@@ -1,8 +1,6 @@
 ﻿namespace EightyDecibel.AsyncNats.Messages
 {
     using System;
-    using System.Buffers;
-    using System.Buffers.Text;
     using System.Collections.Generic;
     using System.Linq;
     using System.Text;