--- a/AsyncNats/Messages/NatsOk.cs
+++ b/AsyncNats/Messages/NatsOk.cs
@@ -12,4 +12,6 @@
             return _instance;
         }
     }
+
+    
 }