Followed your idea @80dB 
Performance was not hit at all apparently.