@israellot 2022-08-29
Ok, I ended up trying another approach to the publish path, not renting memory anymore but writing directly to buffers that rotate. I think it was a breakthrough, the library can now flush twice more messages with no allocations. 
It turns out that on high volume scenarios, the ArrayPool would be exhausted and it would end up allocating anyway. And for small buffers, like a 32 byte message, renting and returning is way more expensive than a simple new[]. 
Take a look and let me know your thoughts. 
I have some ideas for the read path as well, stay tuned.

![Screenshot 2022-08-29 085038](https://user-images.githubusercontent.com/840042/187200324-a1beedf1-7433-4c75-bb02-f13aeb66973d.png)