@israellot 2022-09-01
Added one more proposal here. Inline subscriptions. 
Those are subscriptions that are processed directly in the sequence reader loop, without any extra copying of data. 
This adds huge performance gains for use cases where messages are small. You'll notice on benchmark that as messages get longer, the performance gap diminishes. It can now ingest ~4M messages/sec using this technic. Of course, that's something of a ceiling, as the benchmark just reads a long directly from the payload data, but it gives us an idea of how much overhead can be avoided. 
I added the subscribe with *Unsafe sufix, although I believe it can be made completely safe since the delegate argument is a readonly ref struct ( can't hold reference to it ), and the only concern would be holding a reference to the payload's ReadOnlySequence<byte>, which could be made a Span<byte> for most cases where the payload received is not split among multiple spans. 
Using this, roundtrips benchmarks have almost no allocations going on, besides the async state machine for the occasional awaits.