@80dB 2022-09-06
I'm currently looking through the changes. One thing I noticed iss that while "Inline" is (in some cases much) faster, the "Normal" is slower.  So looking into that a bit.

Old performance:
````
---Roundtrip raw pub 1 sub---
Target flood            0 B     1 pub   1 sub   :       3471,9k msg/s   6,94 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            8 B     1 pub   1 sub   :       3710,7k msg/s   10,39 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            16 B    1 pub   1 sub   :       2773,2k msg/s   10,26 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            32 B    1 pub   1 sub   :       2883,1k msg/s   15,28 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            64 B    1 pub   1 sub   :       2873,6k msg/s   24,43 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            128 B   1 pub   1 sub   :       3664,8k msg/s   54,97 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            256 B   1 pub   1 sub   :       2883,0k msg/s   80,15 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            512 B   1 pub   1 sub   :       2266,2k msg/s   121,02 MB/s Latency: 0,00ms 0,00ms 0,00ms
Target flood            1024 B  1 pub   1 sub   :       1417,1k msg/s   148,51 MB/s Latency: 0,00ms 0,00ms 0,00ms
```` 

New performance:
````
---Roundtrip raw pub 1 sub---
Target flood     (Normal)       0 B     1 pub   1 sub : 2792,0k msg/s   5,58 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       8 B     1 pub   1 sub : 3044,4k msg/s   8,52 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       16 B    1 pub   1 sub : 2896,0k msg/s   10,72 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       32 B    1 pub   1 sub : 3172,5k msg/s   17,13 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       64 B    1 pub   1 sub : 2761,9k msg/s   23,75 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       128 B   1 pub   1 sub : 2898,3k msg/s   43,76 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       256 B   1 pub   1 sub : 2393,9k msg/s   66,79 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       512 B   1 pub   1 sub : 1956,3k msg/s   104,66 MB/s Lat: 0,00ms 0,00ms 0,00ms
Target flood     (Normal)       1024 B  1 pub   1 sub : 1193,0k msg/s   125,03 MB/s Lat: 0,00ms 0,00ms 0,00ms
````

Some discrepancies but it seems that the "old" normal is faster. 

No doubt about Publish though, it's definitely faster.   Did you look at serializing the message in the "WriteSocketAsync" task vs the current NatsPublishBuffer/Channel implementation?