@israellot 2022-09-06
Yes, I did try serializing after dequeue. The issue is mostly for smaller messages. As we can’t guarantee the payload buffer the user supplied will be valid, we need to either copy it or make the user wait for serialization with a task completion. That combined with the overhead of queuing that many messages slows down the entire process.