@israellot 2022-09-06
Regarding the read performance issue, yes, for smaller messages any extra overhead ( like and extra if ) ends up affecting performance. One experiment I did for example, is to drop the rented memory approach for small messages and let GC take care of cleanup. It actually doubles msg/sec, at the expense of more GC, but no explosions. I believe looking at those numbers at face value is not the right way to evaluate though. If you have a system throwing 2M / msgs per second at you in a single topic you will probably want to avoid a 100ms GC pause. My aim was actually to remove overheads, specially allocations, because those ultimately translates into GC pauses.
I guess my point is: if your use case is processing small messages in high volume, inline is the recommended approach. If your messages are bigger or are in sane volumes, it doesn’t matter much which approach you choose, both can handle, inline will offer you less allocations if you can examine the message without copying it. If you need to copy the payload anyway, then it’s irrelevant again which method you choose.