@80dB 2022-09-07
> As we can’t guarantee the payload buffer the user supplied will be valid, we need to either copy it or make the user wait for serialization with a task completion. 

The only time we can't guarantee the payload buffer is when the user supplies his/her own byte[] , in every other instance [object or string] we can (and I suspect object to be the most common payload). Because with string the implicit conversion handles this and with an object the serialization from object to byte[] is also done by us.

The serialization from something to NatsPayload still needs to be done in the publish though, but the copy would only need to happen if "they" supplied their own NatsPayload [that is not based off a string].