@israellot 2022-09-07
That's true. But if you think in terms of (asm operations/message byte), enqueuing and dequeing every message to a channel will necessarily degrade performance for smaller messages just due to the fact these operations add up, more jpms on while loops, allocations to hold elements in the queue, locks, etc. If the message size is big enough, eventually the cost is amortized and the approach won't matter much. Unfortunately, I didn't save the post queue serialization experiment I did, but shouldn't be hard to try it if you want to explore this avenue.