@80dB 2022-09-12
I've played with queuing messages and like you expected, it's performance is lower than this branch. It's faster than the current implementation but slower than your implementation.

I did have a version that implemented a byte[]-pool and BlockingCollection, it was faster with <128b messages [11.2m vs 6.7m on 0-byte PUB's and 7.8m vs 6.1m on 128-byte PUB's], but it's performance tanked at 1024b [and higher] messages. However it's CPU-usage was through the roof [probably due to the lack of proper async-handling]. It does show potential though :)